Budget, Goal minor things fixed!~

+++ ToTe/Goal.h
@@ -18,16 +18,16 @@
 @property(nonatomic)NSString *goal_photo;
 @property(nonatomic)int priority;
 @property(nonatomic)int weeks_met;
-@property(nonatomic)int amount_tosave;
+@property(nonatomic)double amount_tosave;
 @property(nonatomic)NSString* goal_start_date;
 
--(int)InsertGoal:(NSString *)g_title:(NSString *)g_description:(int)g_amount:(NSString *)deadline:(NSString *)g_photo:(int)amount_tosave;
+-(int)InsertGoal:(NSString *)g_title:(NSString *)g_description:(int)g_amount:(NSString *)deadline:(NSString *)g_photo:(double)amount_tosave;
 -(NSString *)ConvertDateFormat:(NSString *)end_date;
 -(NSString *)getCurrentDay;
 -(NSInteger)WeeksBetweenDate:(NSDate *)end_date;
 -(NSMutableArray *)SelectAllGoals;
 -(NSMutableArray *)SelectGoal:(int)g_id;
--(BOOL)UpdateGoal:(NSString *)g_title:(NSString *)g_description:(int)g_amount:(NSString *)deadline:(NSString *)g_photo:(int)amount_tosave:(int)g_id;
+-(BOOL)UpdateGoal:(NSString *)g_title:(NSString *)g_description:(int)g_amount:(NSString *)deadline:(NSString *)g_photo:(double)amount_tosave:(int)g_id;
 -(BOOL)DeleteGoal:(int)g_id;
 -(NSInteger)WeeksBetweenTwoDate:(NSDate *)start_date:(NSDate *)end_date;
 -(NSDate *)StringToDate:(NSString *)strDate;
